Prepare files and add dependecies to write tests
